requirements.txt
annotated-types==0.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyYAML==6.0.3
rich==14.3.2
typing-inspection==0.4.2
typing_extensions==4.15.0
